Use Moon logo instead of Dracula logo for Hyper theme
The Dracula logo is only used at the top of the page and looks a little out of
place as a theme logo. moon.png was used becuase it would be the next in the
cycle if we continued to use the same logos in the same order. This is probably
best for now until we get more logos