Add instructions for vim-plug (#17)
Although vim plug would normally be simple enough and somewhat similar
to vundle, it is important to give it the `as` option to make it install
dracula in it's own 'dracula' directory. This is because the repo is
abnormally named `vim` (that's not a standard in the community) and can
cause conflicts if others do the same.